refactor(ui): deduplicate shared components and improve code quality
Extract DependencyListModel and StepListView as shared components to
eliminate duplicated logic between SetupView/SettingsView and
GenerationProgressView/RefineProgressView. Make PluginType properly
RawRepresentable and clean up alert bindings in PluginLibraryView.
