+++ Foundry/Views/ErrorView.swift
@@ -93,7 +93,8 @@ struct ErrorView: View {
             ToolbarItem(placement: .automatic) {
                 Button("Edit Prompt") {
                     appState.popToRoot()
-                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
+                    Task {
+                        try? await Task.sleep(for: .milliseconds(100))
                         appState.push(.prompt)
                     }
                 }
