+++ Foundry/Views/GenerationProgressView.swift
@@ -34,7 +34,6 @@ struct GenerationProgressView: View {
 
     @State private var pipeline = GenerationPipeline()
     @State private var elapsedSeconds: Int = 0
-    @State private var timer: Timer?
     @State private var completedSteps: Set<Int> = []
 
     var body: some View {
@@ -75,42 +74,12 @@ struct GenerationProgressView: View {
                     }
                 }
 
-                // Step list
-                VStack(spacing: 2) {
-                    ForEach(GenerationStep.allCases, id: \.self) { step in
-                        HStack(spacing: 10) {
-                            stepIndicator(for: step)
-                                .frame(width: 16)
-
-                            Image(systemName: step.icon)
-                                .font(.system(size: 11))
-                                .foregroundStyle(stepColor(for: step))
-                                .frame(width: 16)
-
-                            Text(step.label)
-                                .font(.subheadline)
-                                .foregroundStyle(stepColor(for: step))
-
-                            Spacer()
-
-                            if completedSteps.contains(step.rawValue) {
-                                Text("Done")
-                                    .font(.caption2)
-                                    .foregroundStyle(.tertiary)
-                            }
-                        }
-                        .padding(.vertical, 8)
-                        .padding(.horizontal, 12)
-                        .background(
-                            step == pipeline.currentStep
-                                ? Color.accentColor.opacity(0.06)
-                                : Color.clear,
-                            in: .rect(cornerRadius: 6)
-                        )
-                    }
-                }
-                .padding(4)
-                .background(Color(.controlBackgroundColor).opacity(0.3), in: .rect(cornerRadius: 10))
+                StepListView(
+                    steps: GenerationStep.allCases,
+                    currentStep: pipeline.currentStep,
+                    completedSteps: completedSteps,
+                    buildAttempt: pipeline.buildAttempt
+                )
             }
             .frame(maxWidth: 440)
 
@@ -127,17 +96,18 @@ struct GenerationProgressView: View {
             ToolbarItem(placement: .cancellationAction) {
                 Button("Cancel") {
                     pipeline.cancel()
-                    timer?.invalidate()
                     appState.popToRoot()
                 }
             }
         }
         .onAppear {
-            startTimer()
             pipeline.run(config: config, appState: appState)
         }
-        .onDisappear {
-            timer?.invalidate()
+        .task {
+            while !Task.isCancelled {
+                try? await Task.sleep(for: .seconds(1))
+                elapsedSeconds += 1
+            }
         }
         .onChange(of: pipeline.currentStep) { oldValue, newValue in
             if newValue.rawValue > oldValue.rawValue {
@@ -145,34 +115,6 @@ struct GenerationProgressView: View {
                     _ = completedSteps.insert(oldValue.rawValue)
                 }
             }
-        }
-    }
-
-    @ViewBuilder
-    private func stepIndicator(for step: GenerationStep) -> some View {
-        if completedSteps.contains(step.rawValue) {
-            Image(systemName: "checkmark")
-                .font(.system(size: 9, weight: .bold))
-                .foregroundStyle(.green)
-        } else if step == pipeline.currentStep {
-            ProgressView()
-                .controlSize(.mini)
-        } else if step.rawValue > pipeline.currentStep.rawValue {
-            Circle()
-                .fill(.quaternary)
-                .frame(width: 6, height: 6)
-        } else {
-            EmptyView()
-        }
-    }
-
-    private func stepColor(for step: GenerationStep) -> some ShapeStyle {
-        if completedSteps.contains(step.rawValue) {
-            return .tertiary
-        } else if step == pipeline.currentStep {
-            return .primary
-        } else {
-            return .quaternary
         }
     }
 
@@ -184,14 +126,6 @@ struct GenerationProgressView: View {
         let minutes = elapsedSeconds / 60
         let seconds = elapsedSeconds % 60
         return String(format: "%d:%02d", minutes, seconds)
-    }
-
-    private func startTimer() {
-        timer = Timer.scheduledTimer(withTimeInterval: 1, repeats: true) { _ in
-            Task { @MainActor in
-                elapsedSeconds += 1
-            }
-        }
     }
 }
